test/pyproject.toml
[project]
name = "verifier-agent-test-client"
version = "0.1.0"
description = "Test client for verifier agent service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]
